refactored models in Mock Server
