+++ mock-server/src/models/exchanged-price.model.ts
@@ -1,5 +1,5 @@
 import { Currency } from './currency.model';
-import { Price } from './prise.model';
+import { Price } from './price.model';
 
 export interface ExchangedPrice extends Price {
   originalCurrency: Currency;
