+++ mock-server/src/models/expense.model.ts
@@ -1,6 +1,6 @@
 import { Category } from './category.model';
 import { ExchangedPrice } from './exchanged-price.model';
-import { Price } from './prise.model';
+import { Price } from './price.model';
 
 export interface Expense {
   id?: number;
