+++ mock-server/src/models/expenses-view.model.ts
@@ -1,0 +1,7 @@
+import { Expense } from './expense.model';
+import { Price } from './price.model';
+
+export interface ExpensesView {
+  expenses: Expense[];
+  total?: Price;
+}
